[IMP] mail: turn 'Message/trackingValues' into field
Task-2793280
